listing_service/requirements.txt
fastapi[all]
pydantic
pymongo[srv]
python-dotenv
pytest